+++ changelog.md
@@ -1,5 +1,23 @@
 ## grblHAL changelog
 
+<a name="20250103">20250103
+
+Core:
+
+* Added new handling for spindle spin up and coolant start delays. Setting `$394` is for spindle spin up and `$673` for coolant start.  
+They can be either set to 0 or to a value in the range 0.5 - 20s. The settings are used both for initial delay and for restore delay after cancelling a feed hold or completing a tool change.  
+> [!NOTE]
+> If the spindle supports "at speed" functionality and this is enabled by setting `$340` \(Spindle at speed tolerance\) > 0 then the spin up delay is used as a timeout value before alarm 14 is raised. If `$394` is set to 0 the timeout will default to one minute.
+
+> [!NOTE]
+> Setting `$392` and `$393`, if available, are now only used for spindle spin up delay and coolant start delay respectively when the safety door is closed.
+
+Plugins:
+
+* Laser, coolant: updated for internal naming changes, no functional changes.
+
+---
+
 <a name="20250102">20250102
 
 Core: